Move separate model implementations to separate packages
